Fix long standing issue with area light check
- Properly color based on if the current value is within relevant range
- Switch MathUtil#withinRange to use doubles
- Minor cleanup

+++ src/main/java/com/almuradev/almura/shared/util/MathUtil.java
@@ -58,7 +58,7 @@ public class MathUtil {
      * @param end The end of the range
      * @return True if within the range
      */
-    public static boolean withinRange(float value, float start, float end) {
+    public static boolean withinRange(double value, double start, double end) {
         return value >= start && value <= end;
     }
 }
